Move hymod_state struct to dedicated file.
